\name{Transform}
\alias{TFD}
\title{Transform Catergorical data into usable format}
\usage{
Transform_Fin_Data("C:/Users/vyoma/Desktop/Data.csv")
}
\description{
Transforms the value provided by FinTech which is mostly categorical into usable format."
}
\examples{
Transform_Fin_Data("C:/Users/vyoma/Desktop/Data.csv")
}
